Allow skipping of signature verification when validating Response
Many things are validated in ValidateResponse() after the signature
is verified.  For example, validateAssertion() is called only after
signature verification.  If you have a Response that has been formatted
in any matter (you didn't get the base64 for example), the canonicalized
form may not match the signature anymore.  It would be nice to know that,
barring the signature (and value of 'now'), the Response is otherwise
valid for the given ServiceProvider.